services/auth-service/requirements.txt
flask
flask-cors
mysql-connector-python
bcrypt
pyjwt